Convert TB_WINDOW_MS to double for fractional window support
Changed TB_WINDOW_MS from integer to double (1 → 1.0) to allow
fractional values like 1.4ms, 2.5ms etc.

Updated all pacing formulas to use proper floating-point arithmetic:
- tx_rx_manager.c: DPDK pair TX pacing
- dpdk_external_tx.c: DPDK external TX pacing
- raw_socket_port.c: Raw socket TX pacing (also fixed printf format)

Example rates with TB_WINDOW_MS=1.4:
- Per VL: 1000/1.4 = 714 pkt/s (instead of 1000 pkt/s)
- 70 VLs: 50,000 pkt/s per queue
- Total: ~2.42 Gbps per port (instead of ~3.39 Gbps)

+++ dpdk/src/dpdk_external_tx.c
@@ -286,7 +286,7 @@ int dpdk_ext_tx_worker(void *arg)
     uint16_t total_tb_vl_count = 0;
     for (int t = 0; t < target_count; t++)
         total_tb_vl_count += port_config->targets[t].vl_id_count;
-    uint64_t packets_per_sec = (uint64_t)total_tb_vl_count * TB_PACKETS_PER_VL_PER_WINDOW * 1000 / TB_WINDOW_MS;
+    uint64_t packets_per_sec = (uint64_t)((double)total_tb_vl_count * TB_PACKETS_PER_VL_PER_WINDOW * 1000.0 / TB_WINDOW_MS);
     uint64_t delay_cycles = (packets_per_sec > 0) ? (tsc_hz / packets_per_sec) : tsc_hz;
 #else
     // Hassas hesaplama: rate_mbps -> bytes/sec -> packets/sec -> cycles/packet
